# frontend/packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NookOS API",
    "description": "The open workspace operating system for modern software teams.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/git-credentials": {
      "get": {
        "tags": [
          "crate::routes::gitops"
        ],
        "operationId": "list_git_credentials",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GitCredential"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::routes::gitops"
        ],
        "operationId": "create_git_credential",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGitCredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitCredential"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/git-credentials/{id}": {
      "delete": {
        "tags": [
          "crate::routes::gitops"
        ],
        "operationId": "delete_git_credential",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/nodes/{id}/clone": {
      "post": {
        "tags": [
          "crate::routes::gitops"
        ],
        "operationId": "clone_repo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nodes/{id}/projects": {
      "post": {
        "tags": [
          "crate::routes::gitops"
        ],
        "operationId": "init_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces/{id}/git": {
      "get": {
        "tags": [
          "crate::routes::workspaces"
        ],
        "summary": "Live git status + working-tree diff, relayed from the node.",
        "operationId": "workspace_git_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NodeId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/workspaces/{id}/secrets": {
      "get": {
        "tags": [
          "crate::routes::gitops"
        ],
        "operationId": "list_secrets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceSecret"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{id}/secrets/{name}": {
      "get": {
        "tags": [
          "crate::routes::gitops"
        ],
        "operationId": "get_secret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSecret"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "crate::routes::gitops"
        ],
        "operationId": "put_secret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutSecretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{id}/worktrees": {
      "post": {
        "tags": [
          "crate::routes::gitops"
        ],
        "operationId": "add_worktree",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorktreeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{id}/worktrees/remove": {
      "post": {
        "tags": [
          "crate::routes::gitops"
        ],
        "operationId": "remove_worktree",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveWorktreeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloneRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "credential_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GitCredentialId",
                "description": "Tenant git credential to clone with (private repos over SSH)."
              }
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Directory name; derived from the URL when omitted."
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "CreateGitCredentialRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "generate": {
            "type": "boolean",
            "description": "…or let the server generate an ed25519 keypair."
          },
          "name": {
            "type": "string"
          },
          "private_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "Paste an existing private key (OpenSSH PEM)…"
          }
        }
      },
...
      "GitCredential": {
        "type": "object",
        "description": "A tenant git credential — only the public half is ever returned.",
        "required": [
          "id",
          "tenant_id",
          "name",
          "kind",
          "public_key",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/GitCredentialId"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "public_key": {
            "type": "string"
          },
          "tenant_id": {
            "$ref": "#/components/schemas/TenantId"
          }
        }
      },
      "GitCredentialId": {
        "type": "string",
        "format": "uuid"
      },
      "GitFileStatus": {
        "type": "object",
        "required": [
          "status",
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "Porcelain status code, e.g. \" M\", \"??\", \"A \"."
          }
        }
      },
      "GitStatusResponse": {
        "type": "object",
        "required": [
          "dirty",
          "files",
          "diff"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "diff": {
            "type": "string",
            "description": "Unified diff of the working tree (truncated by the node if huge)."
          },
          "dirty": {
            "type": "boolean"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GitFileStatus"
            }
          }
        }
      },
...
      "InitProjectRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "NodeId": {
        "type": "string",
        "format": "uuid"
      },
...
      "OpResponse": {
        "type": "object",
        "description": "Outcome of a long-running git operation on a node.",
        "required": [
          "ok",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "PutSecretRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
      "RemoveWorktreeRequest": {
        "type": "object",
        "required": [
          "node_id",
          "path"
        ],
        "properties": {
          "node_id": {
            "$ref": "#/components/schemas/NodeId"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "TenantId": {
        "type": "string",
        "format": "uuid"
      },
...
      "WorkspaceSecret": {
        "type": "object",
        "description": "A workspace secret file (e.g. .env). Content only present on single-get.",
        "required": [
          "name",
          "updated_at"
        ],
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "WorktreeRequest": {
        "type": "object",
        "required": [
          "node_id",
          "branch"
        ],
        "properties": {
          "branch": {
            "type": "string"
          },
          "node_id": {
            "$ref": "#/components/schemas/NodeId"
          }
        }
      }
    }
  }
}